fix: rename ShadowNodes to SvgShadowNodes
