+++ tester/harmony/svg/src/main/cpp/SvgShadowNodes.cpp
@@ -8,7 +8,7 @@
  * @generated by codegen project: GenerateShadowNodeCpp.js
  */
 
-#include "ShadowNodes.h"
+#include "SvgShadowNodes.h"
 
 namespace facebook {
 namespace react {
