+++ tester/harmony/svg/src/main/cpp/componentInstances/RNSVGCircleComponentInstance.h
@@ -25,7 +25,7 @@
 #pragma once
 
 #include "RNSVGBaseComponentInstance.h"
-#include "ShadowNodes.h"
+#include "SvgShadowNodes.h"
 
 namespace rnoh {
 namespace svg {
